Revert changes of moving `resolve_path` function. Add 2 new symbols to test the overhead caused by adding new symbols.
