+++ compiler/rustc_span/src/symbol.rs
@@ -557,6 +557,7 @@ symbols! {
         debug_trait_builder,
         debug_tuple,
         debugger_visualizer,
+        debuggervisualizer,
         decl_macro,
         declare_lint_pass,
         decode,
@@ -929,6 +930,7 @@ symbols! {
         native_link_modifiers_verbatim,
         native_link_modifiers_whole_archive,
         natvis_file,
+        natvisfile,
         ne,
         nearbyintf32,
         nearbyintf64,
